Enum file_lock::fd::Mode

pub enum Mode {
    Read,
    Write,
}

Represents a file access mode, e.g. read or write

Variants

Read

Use this to obtain a shared lock, i.e. there may be any amount of readers at the same time.

Write

Used to indicate an exclusive lock, i.e. there may only be one writer at a time.