Function file_lock::fd::unlock

pub fn unlock(fd: RawFd) -> Result<(), Error>

Unlocks the file held by Lock.

In reality, you shouldn't need to call unlock(). As Lock implements the Drop trait, once the Lock reference goes out of scope, unlock() will be called automatically.

For an example, please see the documentation of the Lock structure.