Struct file_lock::filename::Lock

pub struct Lock {
    // some fields omitted
}

A type creating a lock file on demand.

It supports multiple reader, single writer semantics and encodes whether read or write access is required in an interface similar to the one of the RwLock

It will remove the lock file it possibly created in case a lock could be obtained.

Methods

impl Lock

fn new(path: PathBuf, mode: Mode) -> Lock

fn any_lock(&mut self, kind: Kind) -> Result<(), Error>

fn lock(&mut self) -> Result<(), Error>

fn try_lock(&mut self) -> Result<(), Error>

fn unlock(&mut self) -> Result<(), Error>

fn path(&self) -> &PathBuf

fn file(&mut self) -> Option<&mut File>